swagger_server/controllers/branches_controller.py
import connexion
import six

from swagger_server.models.branch import Branch  # noqa: E501
from swagger_server.models.branch_create_request import BranchCreateRequest  # noqa: E501
from swagger_server.models.branch_list import BranchList  # noqa: E501
from swagger_server.models.branch_update_request import BranchUpdateRequest  # noqa: E501
from swagger_server.models.error import Error  # noqa: E501
from swagger_server import util


def branches_branch_id_put(body, branch_id):  # noqa: E501
    """Update branch details

    Update details of an existing branch by ID. # noqa: E501

    :param body: 
    :type body: dict | bytes
    :param branch_id: ID of the branch to update
    :type branch_id: str

    :rtype: Branch
    """
    if connexion.request.is_json:
        body = BranchUpdateRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def branches_delete(branch_id):  # noqa: E501
    """Delete a branch

    Delete a branch by ID. # noqa: E501

    :param branch_id: ID of the branch to delete
    :type branch_id: str

    :rtype: None
    """
    return 'do some magic!'


def branches_get():  # noqa: E501
    """Retrieve all branches

    Get a list of all branches. # noqa: E501


    :rtype: BranchList
    """
    return 'do some magic!'


def branches_post(body):  # noqa: E501
    """Create a new branch

    Add a new branch to the system. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: Branch
    """
    if connexion.request.is_json:
        body = BranchCreateRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'
